view "stream stopped" when you stopped the stream
git-svn-id: https://nuts-radio-a3.googlecode.com/svn/trunk@172 c6d3d5e5-5d5d-ad11-a772-7790f10f8155

+++ source/menu.c
@@ -654,7 +654,7 @@ static void volume_up(struct menu* this)
 */
 static void volume_down(struct menu* this)
 { 
-        if(volume != 255)
+        if((int)volume != 255)
         {
         volume += 5;        
         VsSetVolume((int)volume,(int)volume);
